% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_mito_to_csv}
\alias{batch_mito_to_csv}
\alias{batch_csv_to_fasta}
\title{Batch process MitoFish output}
\usage{
batch_mito_to_csv(file_dir, out_dir, ...)

batch_csv_to_fasta(csv_dir, out_dir, ...)
}
\arguments{
\item{file_dir}{character, the file path to the directory
containing input files.}

\item{out_dir}{character, the file path to the output directory to
write files to. If the directory does not exist, it will be
created.}

\item{...}{additional args passed to the processing function}
}
\value{
list of processed output, in the case of mito_to_csv, or
    the file path to the written files, in the case of
    csv_to_fasta
}
\description{
These functions allow for batch processing MitoFish output. Each
takes a directory full of files.
}
\author{
Matt Espe
}
