% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_extract_ncbi_tabs}
\alias{batch_extract_ncbi_tabs}
\alias{batch_reduce_ncbi}
\title{Batch reduce NCBI tables}
\usage{
batch_extract_ncbi_tabs(
  file_dir,
  files = list.files(file_dir, pattern = "\\\\.txt$", full.names = TRUE, recursive =
    TRUE),
  ...
)

batch_reduce_ncbi(tab_list, accession_filter = c("all", "top"), out_dir, ...)
}
\arguments{
\item{file_dir}{the directory to look for raw NCBI output files}

\item{files}{vector, the full file path to the raw NCBI files}

\item{...}{additional args passed to \code{reduce_nbci}}

\item{tab_list}{a list of extracted tables created by
\code{batch_extract_ncbi_tabs}}

\item{accession_filter}{character, one of "all" to keep all
accessions, or "top" to only keep the top accession.}

\item{out_dir}{optional, if specified, the reduced tables will be
written to CSV in the directory specified. If the directory
does not exist, it will be created.}
}
\value{
list of reducded tables
}
\description{
Convenience function to batch process NCBI tables.
}
\author{
Matt Espe
}
