% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{combine_fasta}
\alias{combine_fasta}
\title{Combine fasta files into one}
\usage{
combine_fasta(
  fasta_dir,
  fasta_pattern,
  max_records,
  new_names = paste0("combined_", fasta_pattern, "_"),
  out_dir,
  fasta_files = list.files(fasta_dir, fasta_pattern, full.names = TRUE)
)
}
\arguments{
\item{fasta_dir}{directory containing fasta files to combine}

\item{fasta_pattern}{the pattern to match for fasta files
(e.g. "NoHit")}

\item{max_records}{optional, maximum number of records for each
file. The total lines in each file will be twice this number
(since records span two lines). If not provided, all files
will be written to a single fasta file}

\item{new_names}{character, the base of the name for the new
combined files}

\item{out_dir}{optional, if not provided combined files will be
written into the same directory as input files.}

\item{fasta_files}{vector of files to combine}
}
\value{
file names of new combined files
}
\description{
Takes a directory full of fasta files, and combines them into a
single one.
}
\author{
Matt Espe
}
