% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_fasta.R
\name{csv_to_fasta}
\alias{csv_to_fasta}
\title{CSV to Fasta}
\usage{
csv_to_fasta(
  csv_file,
  df = read.csv(csv_file),
  outfile = gsub("\\\\.csv$", "_fasta.txt", basename(csv_file)),
  outdir = "."
)
}
\arguments{
\item{csv_file}{file path to the CSV to read in}

\item{df}{data.frame, the results of read.csv}

\item{outfile}{name of output file}

\item{outdir}{directory to write output to. Defaults to current working directory}
}
\description{
Creates a Fasta file from a CSV
}
\author{
Matt Espe
}
