% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ncbi.R
\name{get_queries}
\alias{get_queries}
\title{Get Queries from NCBI result}
\usage{
get_queries(file, lines = readLines(file))
}
\arguments{
\item{file}{character, the name of the file to return queries from}

\item{lines}{individual lines of the output from \code{readLines(file)}}
}
\value{
data.frame with columns query (string) and results (logical)
}
\description{
Convenience function for returning all the queries returned in an
NCBI results file, including those without results
}
\author{
Matt Espe
}
