% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabarcodeUtils-pkg.R
\docType{package}
\name{metabarcodeUtils}
\alias{metabarcodeUtils}
\title{metabarcodeUtils: A package of utilities for the eDNA
metabarcoding workflow}
\description{
metabarcodeUtils provides various utilities for meta-barcoding
eDNA data.
}
