% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mitofish.R
\name{mitofish_to_csv}
\alias{mitofish_to_csv}
\title{Mitofish to CSV}
\usage{
mitofish_to_csv(file, outfile = basename(file), outdir = ".")
}
\arguments{
\item{file}{file path to the xlsx file output from mitofish}

\item{outfile}{basename for output files, defaults to the name of the input file}

\item{outdir}{directory to write CSVs to. Defaults to current working directory}
}
\value{
list, with a data.frame of hits and a data.frame for nohits
}
\description{
Converts from a mitofish XLSX file to two CSVs. The hit CSV
contains with one row per hit, with the sample name designating
the stregth of the hit. The nohit CSV contains just the sequence of number of hits
}
\author{
Matt Espe
}
