% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ncbi.R
\name{reduce_ncbi}
\alias{reduce_ncbi}
\title{Reduce NCBI BLAST results}
\usage{
reduce_ncbi(
  df,
  keep_cols = c("scientific_name", "common_name", "taxid", "query_cover", "per_ident",
    "accession"),
  accession_keep = c("all", "top"),
  query_threshold = 90,
  thresholds = c(97, 80, 0)
)
}
\arguments{
\item{df}{data.frame}

\item{keep_cols}{vector of column names to keep in the output.}

\item{accession_keep}{either "all" to keep all accessions, or
"top" to only keep the accession with the top per_ident}

\item{query_threshold}{the threshold of query_coverage. Below this
value, a string will be returned.}

\item{thresholds}{vector of threshold to cut off at. The last
should be zero.}
}
\value{
list of data.frames of reduced rows
}
\description{
Reduces the results of the NCBI BLASTN results.
}
\author{
Matt Espe
}
