% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ncbi.R
\name{write_ncbi_tabs}
\alias{write_ncbi_tabs}
\title{Write NCBI data to a CSV}
\usage{
write_ncbi_tabs(
  x,
  outdir = ".",
  out_prefix = "NCBI_reduced_",
  out_names = names(x),
  ...
)
}
\arguments{
\item{x}{either a list or data.frame}

\item{outdir}{the directory to write CSVs to}

\item{out_prefix}{the common prefix for all files}

\item{out_names}{the unique name to use for each table}

\item{...}{addtional args passed to write.csv}
}
\value{
the file paths of the CSVs written
}
\description{
Write tables extracted and possibly reduced to CSV files for
further analysis.
}
\author{
Matt Espe
}
